Mental Health Agent/requirements.txt
asyncio
python-dotenv
requests
aiohttp
flask
mysql-connector-python
ollama
numpy
googletrans
langdetect
duckduckgo_search
rich
prompt_toolkit 
nest_asyncio
deep_translator
# transformers
fastapi
uvicorn
openai